more runtests splitting
